Move SlyDataRedactor to messages
